+++ src/main/java/com/mookive/mookive_backend/user/application/dto/response/UserResponse.java
@@ -1,4 +1,4 @@
-package com.mookive.mookive_backend.user.application.dto;
+package com.mookive.mookive_backend.user.application.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
