+++ src/main/java/com/mookive/mookive_backend/user/application/service/UserCreateService.java
@@ -1,6 +1,6 @@
 package com.mookive.mookive_backend.user.application.service;
 
-import com.mookive.mookive_backend.user.application.dto.UserRequest;
+import com.mookive.mookive_backend.user.application.dto.request.UserRequest;
 import com.mookive.mookive_backend.user.application.mapper.UserMapper;
 import com.mookive.mookive_backend.user.domain.entity.User;
 import com.mookive.mookive_backend.user.domain.service.UserSaveService;
